requirements.txt
fastapi
uvicorn[standard]
streamlit
opencv-python-headless
ultralytics
numpy
requests
python-dotenv
python-multipart
